revise workflow for matrix testing CLI mode
